+++ homeassistant/components/recorder/auto_repairs/statistics/schema.py
