+++ tests/components/recorder/auto_repairs/__init__.py
