+++ tests/components/recorder/auto_repairs/statistics/__init__.py
