+++ tests/components/recorder/auto_repairs/statistics/test_duplicates.py
@@ -13,11 +13,11 @@ from sqlalchemy.orm import Session
 
 from homeassistant.components import recorder
 from homeassistant.components.recorder import statistics
-from homeassistant.components.recorder.const import SQLITE_URL_PREFIX
-from homeassistant.components.recorder.repairs.statistics.duplicates import (
+from homeassistant.components.recorder.auto_repairs.statistics.duplicates import (
     delete_statistics_duplicates,
     delete_statistics_meta_duplicates,
 )
+from homeassistant.components.recorder.const import SQLITE_URL_PREFIX
 from homeassistant.components.recorder.statistics import async_add_external_statistics
 from homeassistant.components.recorder.util import session_scope
 from homeassistant.core import HomeAssistant
